requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-heroku==0.3.1
django-restframework==0.0.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.8
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==1.7.1
pyparsing==2.4.7
PyPDF2==1.26.0
python-decouple==3.3
pytz==2020.4
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
